Fix sign failure when data has UTF-8 characters.
The sign function used to branch out when parsing the data to sign.
It tried to parse a plaintext literal U128 value from either bytes
of the string representation or from bytes of the raw value.
The function is actually never used with string data,
it's always used with a byte buffer.
Sometimes Poseidon8 hash produced a value that had bytes representing
valid UTF-8 characters but that was never an actual string representing
a U128 literal. The fix is to never try interpretting the data to sign
as a UTF-8 string.

This commit also includes improvements on error reporting in the
sign function.